refactor: cart reducer 리팩터링
