+++ src/hooks/useCart.js
@@ -1,23 +1,54 @@
+import axios from 'axios';
 import { useDispatch } from 'react-redux';
-import {
-  addCartItemAsync,
-  deleteCartItemAsync,
-  updateItemQuantityAsync,
-} from '../store/cart/cart.actions';
+import { SERVER_PATH, STORAGE_KEY } from '../constants';
+import { getCartItemAsync } from '../store/cart/cart.actions';
+
+const accessToken = JSON.parse(localStorage.getItem(STORAGE_KEY));
 
 const useCart = () => {
   const dispatch = useDispatch();
 
-  const addItem = (id) => {
-    dispatch(addCartItemAsync(id));
+  const addItem = async (id) => {
+    try {
+      await axios.post(`${SERVER_PATH.CARTS}/products/${id}`, null, {
+        headers: {
+          Authorization: `Bearer ${accessToken}`,
+        },
+      });
+      dispatch(getCartItemAsync(accessToken));
+    } catch (error) {
+      alert(error.response.data);
+    }
   };
 
-  const deleteItem = (id) => {
-    dispatch(deleteCartItemAsync(id));
+  const deleteItem = async (id) => {
+    try {
+      await axios.delete(`${SERVER_PATH.CARTS}/products/${id}`, {
+        headers: {
+          Authorization: `Bearer ${accessToken}`,
+        },
+      });
+      dispatch(getCartItemAsync(accessToken));
+    } catch (error) {
+      alert(error.response.data);
+    }
   };
 
-  const updateItemQuantity = (id, quantity) => {
-    dispatch(updateItemQuantityAsync(id, quantity));
+  const updateItemQuantity = async (id, quantity) => {
+    try {
+      await axios.patch(
+        `${SERVER_PATH.CARTS}/products/${id}`,
+        { quantity },
+        {
+          headers: {
+            Authorization: `Bearer ${accessToken}`,
+          },
+        }
+      );
+      dispatch(getCartItemAsync(accessToken));
+    } catch (error) {
+      alert(error.response.data);
+    }
   };
 
   return { addItem, deleteItem, updateItemQuantity };
